Make application exit on OAuthException
If the user encounters a problem with their authentication we will send
the user back to the Landing Page automatically. The script there should
either tell them to log back in or redirect with a clean token. I've
tested this by logging out of Facebook and forcing to be thrown back to
the log in page. More testing is probably required.