Major clean up, fixed issues with deployment
Fixed the issue where some files where not included in the deployment to
the server and had to be added manually. Found a way to add them to the
project. For future reference if a file isn't being exported right click
on it and click include in project.

Removed a trash file and added a new gitignore file.